# doc/contracts/backend-api-v1.openapi.yaml
openapi: 3.1.0
info:
  title: LabRat Backend API
  version: 1.0.0
  description: >-
    Contract-first API for the NestJS modular monolith. All documented
    operations have closed DTOs and authoritative v1 behavior.
    Public Guest identities are restricted to one demo project with read/export
    access. Authenticated Guest writes, AI requests and invitation redemption
    return 403 public_guest_read_only; only GET/HEAD and POST logout are allowed.
    Guest request limits may return 429 too_many_requests on authenticated reads.
servers:
  - url: http://127.0.0.1:8788
    description: Local v1 development server
...
security:
  - cookieAuth: []
paths:
...
  /api/v1/projects/{projectId}:
    parameters:
      - $ref: "#/components/parameters/ProjectId"
    get:
      tags: [Tenancy]
      operationId: getProject
      x-labrat-capability: read
      responses:
        "200":
          description: Full or minimum authorized Project shell.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Tenancy]
      operationId: updateProject
      x-labrat-capability: propose
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProjectRequest"
      responses:
        "200":
          description: Project updated.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ProjectResponse"
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/projects/{projectId}/evidence/retrieve:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    post:
      tags: [Evidence]
      operationId: retrieveAcceptedEvidence
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/RetrieveEvidenceRequest" }
      responses:
        "200":
          description: Bounded retrieval from accepted RegionUnderstandingRevisions.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/EvidenceRetrievalResponse" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
...
  /api/v1/projects/{projectId}/files:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Evidence]
      operationId: listFileObjects
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/FileObjectPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Evidence]
      operationId: uploadFileObject
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              additionalProperties: false
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "200": { $ref: "#/components/responses/FileUploadResult" }
        "201": { $ref: "#/components/responses/FileUploadResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/projects/{projectId}/import-runs:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Evidence]
      operationId: listImportRuns
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/ImportRunPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Evidence]
      operationId: createImportRun
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateImportRunRequest" }
      responses:
        "201": { $ref: "#/components/responses/ImportRunObject" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/projects/{projectId}/source-documents:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Evidence]
      operationId: listSourceDocuments
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/SourceDocumentPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/projects/{projectId}/workbook-review-sessions:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Evidence]
      operationId: listWorkbookReviewSessions
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewSessionPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Evidence]
      operationId: createWorkbookReviewSession
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWorkbookReviewSessionRequest" }
      responses:
        "201": { $ref: "#/components/responses/WorkbookReviewSessionCreated" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/projects/{projectId}/region-understandings:
    parameters: [{ $ref: "#/components/parameters/ProjectId" }]
    get:
      tags: [Evidence]
      operationId: listAcceptedRegionUnderstandings
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/RegionUnderstandingPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/source-documents/{sourceDocumentId}/regions:
    parameters: [{ $ref: "#/components/parameters/SourceDocumentId" }]
    get:
      tags: [Evidence]
      operationId: listSourceDocumentRegions
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/SourceRegionPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/source-documents/{sourceDocumentId}/query:
    parameters: [{ $ref: "#/components/parameters/SourceDocumentId" }]
    post:
      tags: [Evidence]
      operationId: querySourceDocument
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceDocumentQueryRequest" }
      responses:
        "200": { $ref: "#/components/responses/SourceQueryResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/source-documents/{sourceDocumentId}/range:
    parameters: [{ $ref: "#/components/parameters/SourceDocumentId" }]
    post:
      tags: [Evidence]
      operationId: readSourceDocumentRange
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SourceDocumentRangeRequest" }
      responses:
        "200": { $ref: "#/components/responses/SourceRangeResult" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /api/v1/workbook-review-sessions/{sessionId}:
    parameters: [{ $ref: "#/components/parameters/SessionId" }]
    get:
      tags: [Evidence]
      operationId: getWorkbookReviewSession
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewSessionDetail" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Evidence]
      operationId: deleteWorkbookReviewSession
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/DeleteWorkbookReviewSessionRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewSessionDeleted" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/workbook-review-sessions/{sessionId}/regions:
    parameters: [{ $ref: "#/components/parameters/SessionId" }]
    get:
      tags: [Evidence]
      operationId: listWorkbookReviewRegions
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Evidence]
      operationId: createWorkbookReviewRegion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateWorkbookReviewRegionRequest" }
      responses:
        "201": { $ref: "#/components/responses/WorkbookReviewRegionMutation" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/workbook-review-sessions/{sessionId}/regions/{regionId}:
    parameters:
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/RegionId" }
    get:
      tags: [Evidence]
      operationId: getWorkbookReviewRegion
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionObject" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    delete:
      tags: [Evidence]
      operationId: deleteWorkbookReviewRegion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkbookReviewRegionDispositionRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionObject" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/workbook-review-sessions/{sessionId}/regions/{regionId}/interpret:
    parameters:
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/RegionId" }
    post:
      tags: [Evidence]
      operationId: interpretWorkbookReviewRegion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/InterpretWorkbookReviewRegionRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionMutation" }
        "201": { $ref: "#/components/responses/WorkbookReviewRegionMutation" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/workbook-review-sessions/{sessionId}/regions/{regionId}/revisions:
    parameters:
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/RegionId" }
    get:
      tags: [Evidence]
      operationId: listRegionUnderstandingRevisions
      x-labrat-capability: read
      x-labrat-scope: all_experiments
      responses:
        "200": { $ref: "#/components/responses/RegionUnderstandingRevisionPage" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
    post:
      tags: [Evidence]
      operationId: reviseWorkbookReviewRegion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ReviseWorkbookReviewRegionRequest" }
      responses:
        "201": { $ref: "#/components/responses/WorkbookReviewRegionMutation" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationError" }
  /api/v1/workbook-review-sessions/{sessionId}/regions/{regionId}/confirm:
    parameters:
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/RegionId" }
    post:
      tags: [Evidence]
      operationId: confirmWorkbookReviewRegion
      x-labrat-capability: approve
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ConfirmWorkbookReviewRegionRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionConfirmed" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/workbook-review-sessions/{sessionId}/regions/{regionId}/ignore:
    parameters:
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/RegionId" }
    post:
      tags: [Evidence]
      operationId: ignoreWorkbookReviewRegion
      x-labrat-capability: propose
      x-labrat-scope: all_experiments
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/WorkbookReviewRegionDispositionRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionObject" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
  /api/v1/workbook-review-sessions/{sessionId}/regions/{regionId}/link:
    parameters:
      - { $ref: "#/components/parameters/SessionId" }
      - { $ref: "#/components/parameters/RegionId" }
    post:
      tags: [Evidence]
      operationId: linkWorkbookReviewRegion
      x-labrat-capability: approve
      x-labrat-scope: all_experiments
      description: >-
        Links a confirmed region to a data kind and, when it resolves, an experiment,
        without creating an extraction template or version. The region then counts
        as linked workbook data of that kind for the Experiment Browser and charts,
        exactly like an applied template match.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/LinkWorkbookReviewRegionRequest" }
      responses:
        "200": { $ref: "#/components/responses/WorkbookReviewRegionLinked" }
        "400": { $ref: "#/components/responses/ValidationError" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
...
  /api/v1/source-documents/{sourceDocumentId}/cell-classes:
    get:
      tags:
        - Evidence
      operationId: getSourceCellClasses
      parameters:
        - name: sourceDocumentId
          in: path
          required: true
          schema:
            $ref: "#/components/schemas/OpaqueId"
        - name: sheetName
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 200
        - name: range
          in: query
          required: true
          schema:
            type: string
            minLength: 1
            maxLength: 80
      responses:
        "200":
          description: Success.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SourceCellClassesResponse"
        "400":
          $ref: "#/components/responses/ValidationError"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          $ref: "#/components/responses/NotFound"
        "409":
          $ref: "#/components/responses/Conflict"
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: labrat_session
  parameters:
...
    ProjectId:
      name: projectId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
...
    SourceDocumentId:
      name: sourceDocumentId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
    SessionId:
      name: sessionId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
    RegionId:
      name: regionId
      in: path
      required: true
      schema: { $ref: "#/components/schemas/OpaqueId" }
...
  responses:
    Unauthorized:
      description: Authentication is missing or expired.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Forbidden:
      description: The actor lacks the required capability.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    NotFound:
      description: Resource not found in the authorized scope.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    Conflict:
      description: Version, state, stale-head or idempotency conflict.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
    ValidationError:
      description: Request failed bounded validation.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ErrorResponse" }
...
    FileObjectPage:
      description: Project file evidence.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FileObjectPage" }
    FileUploadResult:
      description: A newly stored file or an idempotently reused file.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/FileUploadResponse" }
    ImportRunPage:
      description: Project import scans.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ImportRunPage" }
    ImportRunObject:
      description: Import scan created.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/ImportRunResponse" }
    SourceDocumentPage:
      description: Immutable source-document indexes for the project.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SourceDocumentPage" }
    WorkbookReviewSessionPage:
      description: Active workbook review sessions.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WorkbookReviewSessionPage" }
    WorkbookReviewSessionCreated:
      description: Workbook review session and its deterministic candidate regions.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/WorkbookReviewSessionCreateResponse" }
    RegionUnderstandingPage:
      description: Accepted region-understanding decisions.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/RegionUnderstandingPage" }
    SourceRegionPage:
      description: Indexed regions in one SourceDocument.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SourceRegionPage" }
    SourceQueryResult:
      description: Bounded SourceDocument search results.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SourceQueryResponse" }
    SourceRangeResult:
      description: Bounded source cells with both flat cells and row-major compatibility views.
      content:
        application/json:
          schema: { $ref: "#/components/schemas/SourceRangeResponse" }
...
  schemas:
...
    UpdateProjectRequest:
      type: object
      additionalProperties: false
      minProperties: 1
      properties:
        name: { type: string, minLength: 1, maxLength: 300 }
        description: { type: string, maxLength: 5000 }
        status: { type: string, enum: [active, archived] }
    ProjectResponse:
      type: object
      additionalProperties: false
      required: [project]
      properties:
        project: { $ref: "#/components/schemas/Project" }
...
    RetrieveEvidenceRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query: { type: string, minLength: 1, maxLength: 2000 }
        includePreview: { type: boolean, default: true }
        includeUnconfirmedSuggestions: { type: boolean, default: false }
    EvidenceRetrievalResponse:
      type: object
      additionalProperties: false
      required: [schemaVersion, projectId, query, mode, planner, toolTrace, results, suggestions, clarification]
      properties:
        schemaVersion: { type: string, const: labrat.evidenceRetrieval.toolAgent.v1 }
        projectId: { $ref: "#/components/schemas/OpaqueId" }
        query: { type: string }
        mode: { type: string, const: tool_agent }
        planner:
          type: object
          additionalProperties: false
          required: [provider, fallbackUsed]
          properties:
            provider: { type: string }
            fallbackUsed: { type: boolean }
        toolTrace: { type: array, items: { $ref: "#/components/schemas/JsonObject" } }
        results: { type: array, items: { $ref: "#/components/schemas/JsonObject" } }
        suggestions: { type: array, items: { $ref: "#/components/schemas/JsonObject" } }
        clarification:
          oneOf:
            - { $ref: "#/components/schemas/JsonObject" }
            - type: "null"
...
    CreateImportRunRequest:
      type: object
      additionalProperties: false
      required: [fileObjectId]
      properties:
        fileObjectId: { $ref: "#/components/schemas/OpaqueId" }
...
    SourceDocumentQueryRequest:
      type: object
      additionalProperties: false
      required: [query]
      properties:
        query: { type: string, minLength: 1, maxLength: 500 }
        limit: { type: integer, minimum: 1, maximum: 200 }
...
    SourceDocumentRangeRequest:
      type: object
      additionalProperties: false
      required: [sheetName, range]
      properties:
        sheetName: { type: string, minLength: 1, maxLength: 200 }
        range: { type: string, minLength: 2, maxLength: 80 }
        maxCells: { type: integer, minimum: 1, maximum: 10000 }
...
    CreateWorkbookReviewSessionRequest:
      type: object
      additionalProperties: false
      properties:
        fileObjectId: { $ref: "#/components/schemas/OpaqueId" }
        sourceDocumentId: { $ref: "#/components/schemas/OpaqueId" }
      oneOf:
        - required: [fileObjectId]
          not: { required: [sourceDocumentId] }
        - required: [sourceDocumentId]
          not: { required: [fileObjectId] }
...